# Migrated hardshrink() to ATen and deprecated nn.Hardshrink()