--- a/aten/doc/Functions.h
+++ b/aten/doc/Functions.h
@@ -411,12 +411,6 @@ static inline Tensor & glu_forward_out(Tensor & output, const Tensor & self, int
 static inline Tensor glu_forward(const Tensor & self, int64_t dim);
 static inline Tensor & glu_backward_out(Tensor & grad_input, const Tensor & grad_output, const Tensor & self, int64_t dim);
 static inline Tensor glu_backward(const Tensor & grad_output, const Tensor & self, int64_t dim);
-static inline Tensor & hardshrink_out(Tensor & output, const Tensor & self, Scalar lambd=0.5);
-static inline Tensor hardshrink(const Tensor & self, Scalar lambd=0.5);
-static inline Tensor & hardshrink_forward_out(Tensor & output, const Tensor & self, Scalar lambd);
-static inline Tensor hardshrink_forward(const Tensor & self, Scalar lambd);
-static inline Tensor & hardshrink_backward_out(Tensor & grad_input, const Tensor & grad_output, const Tensor & self, Scalar lambd);
-static inline Tensor hardshrink_backward(const Tensor & grad_output, const Tensor & self, Scalar lambd);
 static inline Tensor & hardtanh_out(Tensor & output, const Tensor & self, Scalar min_val=-1, Scalar max_val=1);
 static inline Tensor hardtanh(const Tensor & self, Scalar min_val=-1, Scalar max_val=1);
 static inline Tensor & hardtanh_forward_out(Tensor & output, const Tensor & self, Scalar min_val, Scalar max_val);
@@ -2007,24 +2001,6 @@ static inline Tensor & glu_backward_out(Tensor & grad_input, const Tensor & grad
 }
 static inline Tensor glu_backward(const Tensor & grad_output, const Tensor & self, int64_t dim) {
     return infer_type(self).glu_backward(grad_output, self, dim);
-}
-static inline Tensor & hardshrink_out(Tensor & output, const Tensor & self, Scalar lambd) {
-    return infer_type(self).hardshrink_out(output, self, lambd);
-}
-static inline Tensor hardshrink(const Tensor & self, Scalar lambd) {
-    return infer_type(self).hardshrink(self, lambd);
-}
-static inline Tensor & hardshrink_forward_out(Tensor & output, const Tensor & self, Scalar lambd) {
-    return infer_type(self).hardshrink_forward_out(output, self, lambd);
-}
-static inline Tensor hardshrink_forward(const Tensor & self, Scalar lambd) {
-    return infer_type(self).hardshrink_forward(self, lambd);
-}
-static inline Tensor & hardshrink_backward_out(Tensor & grad_input, const Tensor & grad_output, const Tensor & self, Scalar lambd) {
-    return infer_type(self).hardshrink_backward_out(grad_input, grad_output, self, lambd);
-}
-static inline Tensor hardshrink_backward(const Tensor & grad_output, const Tensor & self, Scalar lambd) {
-    return infer_type(self).hardshrink_backward(grad_output, self, lambd);
 }
 static inline Tensor & hardtanh_out(Tensor & output, const Tensor & self, Scalar min_val, Scalar max_val) {
     return infer_type(self).hardtanh_out(output, self, min_val, max_val);