--- a/aten/src/ATen/native/cuda/Activation.cu
+++ b/aten/src/ATen/native/cuda/Activation.cu
@@ -17,7 +17,7 @@ void hardshrink_cuda_kernel(const Tensor& self, Tensor& out_tensor, scalar_t* la
       scalar_t& self_val,
       scalar_t& out_tensor_val,
       bool early_exit) {
-        out_tensor_val = (self_val >= -*lambd && self_val <= *lambd) ? ScalarConvert<double, scalar_t>::to(0.0) : self_val;
+        out_tensor_val = (self_val >= -*lambd && self_val <= *lambd) ? scalar_cast<scalar_t>(0) : self_val;
   });
 }
 
@@ -31,13 +31,13 @@ void hardshrink_backward_cuda_kernel(Tensor& out_tensor, scalar_t* lambd, const 
       scalar_t& self_val,
       scalar_t& grad_val,
       scalar_t& out_tensor_val) {
-        out_tensor_val = (self_val >= -*lambd && self_val <= *lambd) ? ScalarConvert<double, scalar_t>::to(0.0) : grad_val;
+        out_tensor_val = (self_val >= -*lambd && self_val <= *lambd) ? scalar_cast<scalar_t>(0) : grad_val;
   });
 }
 
 Tensor hardshrink_cuda(const Tensor & self, Scalar lambd) {
   auto lambd_tensor = lambd.toTensor().toType(self.type().scalarType()).toBackend(self.is_cuda() ? Backend::CUDA : Backend::CPU);
-  auto out_tensor = at::zeros_like(self);
+  auto out_tensor = at::empty_like(self);
   AT_DISPATCH_FLOATING_TYPES_AND_HALF(self.type(), "hardshrink_cuda", [&] {
     using cuda_scalar_t = cuda::into_type<scalar_t>;
     hardshrink_cuda_kernel<cuda_scalar_t>(self, out_tensor, lambd_tensor.data<cuda_scalar_t>());
@@ -47,7 +47,7 @@ Tensor hardshrink_cuda(const Tensor & self, Scalar lambd) {
 
 Tensor hardshrink_backward_cuda(const Tensor & grad, const Tensor & self, Scalar lambd) {
   auto lambd_tensor = lambd.toTensor().toType(self.type().scalarType()).toBackend(self.is_cuda() ? Backend::CUDA : Backend::CPU);
-  auto out_tensor = at::zeros_like(grad);
+  auto out_tensor = at::empty_like(grad);
   AT_DISPATCH_FLOATING_TYPES_AND_HALF(self.type(), "hardshrink_backward_cuda", [&] {
     using cuda_scalar_t = cuda::into_type<scalar_t>;
     hardshrink_backward_cuda_kernel<cuda_scalar_t>(out_tensor, lambd_tensor.data<cuda_scalar_t>(), self, grad);