--- a/aten/src/THNN/init.cpp
+++ b/aten/src/THNN/init.cpp
@@ -89,9 +89,6 @@
 #include "generic/ELU.c"
 #include "THGenerateFloatTypes.h"
 
-#include "generic/HardShrink.c"
-#include "THGenerateFloatTypes.h"
-
 #include "generic/HardTanh.c"
 #include "THGenerateFloatTypes.h"