## Summary:
1. Added hardshrink() to ATen (CPU + GPU)
2. Removed nn.Hardshrink(), but keeping file "aten/src/THNN/generic/HardShrink.c"
3. Reusing nn.Hardshrink() tests in test_nn and including CUDA tests. Not sure test_nn is a good place for hardshrink() since it is no longer under nn
4. Added tests in test_torch

This replaces #7695, fixes #4154

## Future work:
1. not supporting default lambd=0.5 in torch.hardshrink() due to a Scalar bug #8484


## Performance:

### CPU forward:

Previous impl:
```python
large_data = torch.zeros(1000, 1000).fill_(0.3).requires_grad_()
def origfn(data):
  f = torch.nn.Hardshrink(0.3)
  large_out = f(data)
%timeit origfn(large_data) 
####################
100 loops, best of 3: 2.15 ms per loop
```

Current impl:
```python
large_data = torch.zeros(1000, 1000).fill_(0.3).requires_grad_()
def newfn(data):
    large_out = data.hardshrink(0.3)
%timeit newfn(large_data)
####################
100 loops, best of 3: 2.62 ms per loop
```

### CPU backward:

Previous impl:
```python
large_data = torch.zeros(1000, 1000).fill_(0.3).requires_grad_()
def origfn_backward(data):
    f = torch.nn.Hardshrink(0.3)
    large_out = f(data)
    large_out.sum().backward()
%timeit origfn_backward(large_data)
####################
100 loops, best of 3: 5.2 ms per loop
```

Current impl:
```python
large_data = torch.zeros(1000, 1000).fill_(0.3).requires_grad_()
def newfn_backward(data):
    large_out = data.hardshrink(0.3)
    large_out.sum().backward()
%timeit newfn_backward(large_data)
####################
100 loops, best of 3: 6.47 ms per loop
```

### CUDA:

Current impl:
```python
large_data = torch.zeros(1000, 1000, device=cuda).fill_(0.3)
%timeit data.hardshrink(0.3)
####################
10000 loops, best of 3: 97.1 µs per loop
```

Benchmark:
```python
large_data = torch.zeros(1000, 1000, device=cuda).fill_(0.3)
%timeit data.mul_(2)
####################
10000 loops, best of 3: 51.6 µs per loop
```