@weiyangfb 2018-06-06
The default parameter lambda=0.5 is still not working. Currently I declared it at `native_functions.yaml`, maybe more work is needed? Also, where should I add warning messages for the deprecation of nn.Hardshrink?