@ezyang 2018-06-07
Default should work. Can you push with the test uncommented so we can see the failure?

Re deprecation, cc @soumith and @colesbury, but it won't be easy to "deprecate" the `legacy_nn` codepath in a reasonable way. Maybe best to drop it.