@weiyangfb 2018-06-08
Now the failure should appears:

```
00:21:26 ======================================================================
00:21:26 FAIL: test_hardshrink (test_torch.TestTorch)
00:21:26 ----------------------------------------------------------------------
00:21:26 Traceback (most recent call last):
00:21:26   File "/var/lib/jenkins/workspace/test/test_torch.py", line 5592, in test_hardshrink
00:21:26     self.assertEqual(torch.tensor([1, 0, 0, 0.6]).view(2, 2), data.hardshrink())
00:21:26   File "/var/lib/jenkins/workspace/test/common.py", line 307, in assertEqual
00:21:26     assertTensorsEqual(x, y)
00:21:26   File "/var/lib/jenkins/workspace/test/common.py", line 299, in assertTensorsEqual
00:21:26     self.assertLessEqual(max_err, prec, message)
00:21:26 AssertionError: tensor(0.5000) not less than or equal to 1e-05
00:21:26 
00:21:26 ----------------------------------------------------------------------
00:21:26 lambd = 0.3
00:21:26 lambd = 0.5
00:21:26 lambd = 0
00:21:26 Ran 508 tests in 29.749s
00:21:26 
00:21:26 FAILED (failures=1, skipped=123)
00:21:26 Traceback (most recent call last):
00:21:26   File "test/run_test.py", line 344, in <module>
00:21:26     main()
00:21:26   File "test/run_test.py", line 336, in main
00:21:26     raise RuntimeError(message)
00:21:26 RuntimeError: test_sparse failed!
```