@weiyangfb 2018-06-12
For reasons I don't understand, this PR also allows nn.Hardshrink() to work in both of CPU and CUDA. For instance:

```
cuda = torch.device('cuda')
data = torch.zeros(2,2, device=cuda).fill_(0.3)
f = torch.nn.Hardshrink(0.3)
f(data)
-------------
tensor([[ 0.,  0.],
        [ 0.,  0.]], device='cuda:0')
```

Is nn.Hardshrink() borrowing the implementation from torch.hardshrink()? Did I miss something?