@ezyang 2018-06-13
> Is nn.Hardshrink() borrowing the implementation from torch.hardshrink()? Did I miss something?

That wouldn't surprise me! Have you checked the code?