@ngimel 2018-06-13
nn.Hardshrink is using def from functional.py, and you've told functional.py to call torch.hardshrink. All is good, I think.