@weiyangfb 2018-06-14
@ezyang @ngimel You got it! Should we keep this for now? Or should we just remove nn.Hardshrink() from functional.py? Keeping it allows us to reuse tests in test_nn.py though.