@weiyangfb 2018-06-14
@pytorchbot retest this please