@weiyangfb 2018-06-14
caffe CI test failures look not related, I will merge this PR